# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flexi Route Logistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [awaiting_payment, under_review, confirmed, rejected]
        - name: shipmentId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record for a shipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment (submit TXID or admin review)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/proof:
    post:
      operationId: uploadPaymentProof
      tags: [payments]
      summary: Upload payment screenshot proof
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProofUpload"
      responses:
        "200":
          description: Proof uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # --- WALLETS ---
  /wallets:
    get:
      operationId: getWallets
      tags: [payments]
      summary: Get crypto wallet addresses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet addresses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletAddresses"

  # --- DRIVERS ---
...
  /pricing/{id}:
    patch:
      operationId: updatePricing
      tags: [pricing]
      summary: Update pricing (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingUpdate"
      responses:
        "200":
          description: Updated pricing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pricing"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProofUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        notes:
          type: string

    # --- TRACKING SCHEMAS ---
...
    Payment:
      type: object
      required: [id, shipmentId, amount, currency, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          enum: [BTC, ETH, USDT_TRC20, USDT_ERC20, USDC, LTC]
        walletAddress:
          type: string
        txid:
          type: ["string", "null"]
        paymentProofUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [awaiting_payment, under_review, confirmed, rejected]
        adminNotes:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [shipmentId, amount, currency]
      properties:
        shipmentId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          enum: [BTC, ETH, USDT_TRC20, USDT_ERC20, USDC, LTC]
    PaymentUpdate:
      type: object
      properties:
        txid:
          type: string
        status:
          type: string
          enum: [under_review, confirmed, rejected]
        adminNotes:
          type: string
    PaymentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    WalletAddresses:
      type: object
      required: [BTC, ETH, USDT_TRC20, USDT_ERC20, USDC, LTC]
      properties:
        BTC:
          type: string
        ETH:
          type: string
        USDT_TRC20:
          type: string
        USDT_ERC20:
          type: string
        USDC:
          type: string
        LTC:
          type: string

    # --- DRIVER SCHEMAS ---
...
    Pricing:
      type: object
      required: [id, serviceType, basePriceUsd, pricePerKg, createdAt]
      properties:
        id:
          type: integer
        serviceType:
          type: string
          enum: [standard, express, overnight, freight]
        serviceLabel:
          type: string
        basePriceUsd:
          type: number
        pricePerKg:
          type: number
        estimatedDays:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    PricingUpdate:
      type: object
      properties:
        basePriceUsd:
          type: number
        pricePerKg:
          type: number
        estimatedDays:
          type: string
        description:
          type: string